misc battle cleanup

+++ include/functions.h
@@ -451,7 +451,7 @@ void set_animation_rate(s32 actorID, s32 partIndex, f32 rate);
 void set_model_flags(u16, s32, s32);
 void init_enter_world_shared(void);
 s16 update_enter_map_screen_overlay(s16* progress);
-s32 update_exit_map_screen_overlay(s16* progress);
+s16 update_exit_map_screen_overlay(s16* progress);
 void set_screen_overlay_params_front(u8, f32);
 void set_screen_overlay_params_back(u8, f32);
 void set_screen_overlay_alpha(s32, f32);
